+++ docs/asciidoc/modules/ROOT/partials/periodic-commit.adoc
@@ -26,3 +26,175 @@ call apoc.periodic.commit(
 | Updates | Executions
 | 2000000 | 200
 |===
+
+You can also retrieve the link:https://www.javadoc.io/doc/org.neo4j/neo4j-kernel/2.3.12/org/neo4j/graphdb/event/TransactionData.html[Transaction data]
+from your query with the config `{extractData: true}` to retrieve useful infos.
+For example, from a dataset `UNWIND range(1, 3) AS id CREATE (n:NodeCommit {id:id})`:
+
+[source,cypher]
+----
+CALL apoc.periodic.commit(
+    "MATCH (p:NodeCommit) WHERE NOT p:Processed WITH p LIMIT $limit CREATE (n:Another {alpha: 'beta'}) SET p:Processed RETURN count(*)", 
+    {limit: 2}, 
+    {extractData: true}) YIELD data RETURN data
+----
+
+.Results
+[opts="header"]
+|===
+| data
+| {
+"metaData": {
+
+    },
+    "assignedLabels": {
+        "Processed": [{
+                "identity": 50001,
+                "labels": [
+                    "NodeCommit",
+                    "Processed"
+                ],
+                "properties": {
+                    "id": 1
+                }
+            },
+            {
+                "identity": 50002,
+                "labels": [
+                    "NodeCommit",
+                    "Processed"
+                ],
+                "properties": {
+                    "id": 2
+                }
+            },
+            {
+                "identity": 50003,
+                "labels": [
+                    "NodeCommit",
+                    "Processed"
+                ],
+                "properties": {
+                    "id": 3
+                }
+            }
+        ],
+        "Another": [{
+                "identity": 50004,
+                "labels": [
+                    "Another"
+                ],
+                "properties": {
+                    "alpha": "beta"
+                }
+            },
+            {
+                "identity": 50005,
+                "labels": [
+                    "Another"
+                ],
+                "properties": {
+                    "alpha": "beta"
+                }
+            },
+            {
+                "identity": 50006,
+                "labels": [
+                    "Another"
+                ],
+                "properties": {
+                    "alpha": "beta"
+                }
+            }
+        ]
+    },
+    "removedLabels": {
+
+    },
+    "assignedRelationshipProperties": {
+
+    },
+    "deletedRelationships": [],
+    "deletedNodes": [],
+    "createdRelationships": [],
+    "removedNodeProperties": {
+
+    },
+    "removedRelationshipProperties": {
+
+    },
+    "assignedNodeProperties": {
+        "alpha": [{
+                "node": {
+                    "identity": 50004,
+                    "labels": [
+                        "Another"
+                    ],
+                    "properties": {
+                        "alpha": "beta"
+                    }
+                },
+                "new": "beta",
+                "key": "alpha",
+                "old": null
+            },
+            {
+                "node": {
+                    "identity": 50005,
+                    "labels": [
+                        "Another"
+                    ],
+                    "properties": {
+                        "alpha": "beta"
+                    }
+                },
+                "new": "beta",
+                "key": "alpha",
+                "old": null
+            },
+            {
+                "node": {
+                    "identity": 50006,
+                    "labels": [
+                        "Another"
+                    ],
+                    "properties": {
+                        "alpha": "beta"
+                    }
+                },
+                "new": "beta",
+                "key": "alpha",
+                "old": null
+            }
+        ]
+    },
+    "createdNodes": [{
+            "identity": 50004,
+            "labels": [
+                "Another"
+            ],
+            "properties": {
+                "alpha": "beta"
+            }
+        },
+        {
+            "identity": 50005,
+            "labels": [
+                "Another"
+            ],
+            "properties": {
+                "alpha": "beta"
+            }
+        },
+        {
+            "identity": 50006,
+            "labels": [
+                "Another"
+            ],
+            "properties": {
+                "alpha": "beta"
+            }
+        }
+    ]
+}
+|===
